delete old util functions, add util functions to transform localstorage objects to formData and to get the surveys as formData
